+++ src/main/java/cart/cartitem/exception/CartItemException.java
@@ -1,0 +1,18 @@
+package cart.cartitem.exception;
+
+import cart.common.execption.BaseException;
+import cart.common.execption.BaseExceptionType;
+
+public class CartItemException extends BaseException {
+
+    private final CartItemExceptionType cartItemExceptionType;
+
+    public CartItemException(CartItemExceptionType cartItemExceptionType) {
+        this.cartItemExceptionType = cartItemExceptionType;
+    }
+
+    @Override
+    public BaseExceptionType exceptionType() {
+        return cartItemExceptionType;
+    }
+}
