+++ src/main/java/cart/cartitem/presentation/CartItemApiController.java
@@ -1,0 +1,62 @@
+package cart.cartitem.presentation;
+
+import cart.cartitem.application.CartItemService;
+import cart.cartitem.presentation.dto.AddCartItemRequest;
+import cart.cartitem.presentation.dto.CartItemResponse;
+import cart.cartitem.presentation.dto.UpdateCartItemQuantityRequest;
+import cart.member.domain.Member;
+import java.net.URI;
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/cart-items")
+public class CartItemApiController {
+
+    private final CartItemService cartItemService;
+
+    public CartItemApiController(CartItemService cartItemService) {
+        this.cartItemService = cartItemService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<CartItemResponse>> showCartItems(Member member) {
+        return ResponseEntity.ok(cartItemService.findByMember(member));
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> addCartItems(
+            Member member,
+            @RequestBody AddCartItemRequest request
+    ) {
+        Long cartItemId = cartItemService.add(request.toCommand(member));
+
+        return ResponseEntity.created(URI.create("/cart-items/" + cartItemId)).build();
+    }
+
+    @PatchMapping("/{id}")
+    public ResponseEntity<Void> updateCartItemQuantity(
+            @PathVariable Long id,
+            Member member,
+            @RequestBody UpdateCartItemQuantityRequest request
+    ) {
+        cartItemService.updateQuantity(request.toCommand(id, member));
+
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> removeCartItems(Member member, @PathVariable Long id) {
+        cartItemService.remove(member, id);
+
+        return ResponseEntity.noContent().build();
+    }
+}
