+++ src/main/java/cart/common/config/WebMvcConfig.java
@@ -1,0 +1,39 @@
+package cart.common.config;
+
+import cart.member.infrastructure.persistence.dao.MemberDao;
+import cart.member.presentation.MemberArgumentResolver;
+import java.util.List;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+    private final MemberDao memberDao;
+
+    public WebMvcConfig(MemberDao memberDao) {
+        this.memberDao = memberDao;
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new MemberArgumentResolver(memberDao));
+    }
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:3000", "https://n0eyes.github.io/react-shopping-cart/")
+                .allowedMethods(
+                        HttpMethod.GET.name(),
+                        HttpMethod.POST.name(),
+                        HttpMethod.PUT.name(),
+                        HttpMethod.DELETE.name(),
+                        HttpMethod.PATCH.name()
+                )
+                .allowCredentials(true)
+                .exposedHeaders("*");
+    }
+}
