+++ src/main/java/cart/common/execption/BaseException.java
@@ -1,0 +1,6 @@
+package cart.common.execption;
+
+public abstract class BaseException extends RuntimeException {
+
+    public abstract BaseExceptionType exceptionType();
+}
