+++ src/main/java/cart/common/execption/BaseExceptionType.java
@@ -1,0 +1,12 @@
+package cart.common.execption;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseExceptionType {
+
+    int errorCode();
+
+    HttpStatus httpStatus();
+
+    String errorMessage();
+}
