+++ src/main/java/cart/common/execption/ExceptionResponse.java
@@ -1,0 +1,28 @@
+package cart.common.execption;
+
+public class ExceptionResponse {
+
+    private final int code;
+    private final String message;
+
+    public ExceptionResponse(int code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    @Override
+    public String toString() {
+        return "ExceptionResponse{" +
+                "code=" + code +
+                ", message='" + message + '\'' +
+                '}';
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
