--- src/main/java/cart/dto/CartItemQuantityUpdateRequest.java
@@ -1,16 +1,0 @@
-package cart.dto;
-
-public class CartItemQuantityUpdateRequest {
-    private int quantity;
-
-    public CartItemQuantityUpdateRequest() {
-    }
-
-    public CartItemQuantityUpdateRequest(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-}
