--- src/main/java/cart/exception/AuthenticationException.java
@@ -1,4 +1,0 @@
-package cart.exception;
-
-public class AuthenticationException extends RuntimeException {
-}
