--- src/main/java/cart/exception/CartItemException.java
@@ -1,16 +1,0 @@
-package cart.exception;
-
-import cart.domain.CartItem;
-import cart.domain.Member;
-
-public class CartItemException extends RuntimeException {
-    public CartItemException(String message) {
-        super(message);
-    }
-
-    public static class IllegalMember extends CartItemException {
-        public IllegalMember(CartItem cartItem, Member member) {
-            super("Illegal member attempts to cart; cartItemId=" + cartItem.getId() + ", memberId=" + member.getId());
-        }
-    }
-}
