+++ src/main/java/cart/member/domain/Member.java
@@ -1,0 +1,29 @@
+package cart.member.domain;
+
+public class Member {
+    private Long id;
+    private String email;
+    private String password;
+
+    public Member(Long id, String email, String password) {
+        this.id = id;
+        this.email = email;
+        this.password = password;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public boolean checkPassword(String password) {
+        return this.password.equals(password);
+    }
+}
