+++ src/main/java/cart/member/exception/MemberException.java
@@ -1,0 +1,19 @@
+package cart.member.exception;
+
+import cart.common.execption.BaseException;
+import cart.common.execption.BaseExceptionType;
+
+public class MemberException extends BaseException {
+
+    private final MemberExceptionType memberExceptionType;
+
+    public MemberException(MemberExceptionType memberExceptionType) {
+        this.memberExceptionType = memberExceptionType;
+    }
+
+    @Override
+    public BaseExceptionType exceptionType() {
+        return memberExceptionType;
+    }
+}
+
