+++ src/main/java/cart/member/exception/MemberExceptionType.java
@@ -1,0 +1,40 @@
+package cart.member.exception;
+
+import cart.common.execption.BaseExceptionType;
+import org.springframework.http.HttpStatus;
+
+public enum MemberExceptionType implements BaseExceptionType {
+
+    MISMATCH_AUTHENTICATION(
+            300,
+            HttpStatus.UNAUTHORIZED,
+            "인증정보가 잘못되었습니다."
+    ),
+    ;
+
+    private final int errorCode;
+    private final HttpStatus httpStatus;
+    private final String errorMessage;
+
+    MemberExceptionType(int errorCode, HttpStatus httpStatus, String errorMessage) {
+        this.errorCode = errorCode;
+        this.httpStatus = httpStatus;
+        this.errorMessage = errorMessage;
+    }
+
+    @Override
+    public int errorCode() {
+        return errorCode;
+    }
+
+    @Override
+    public HttpStatus httpStatus() {
+        return httpStatus;
+    }
+
+    @Override
+    public String errorMessage() {
+        return errorMessage;
+    }
+}
+
