+++ src/main/java/cart/member/presentation/MemberArgumentResolver.java
@@ -1,0 +1,59 @@
+package cart.member.presentation;
+
+import static cart.member.exception.MemberExceptionType.MISMATCH_AUTHENTICATION;
+
+import cart.member.domain.Member;
+import cart.member.exception.MemberException;
+import cart.member.infrastructure.persistence.dao.MemberDao;
+import org.apache.tomcat.util.codec.binary.Base64;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.HttpHeaders;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+public class MemberArgumentResolver implements HandlerMethodArgumentResolver {
+    private final MemberDao memberDao;
+
+    public MemberArgumentResolver(MemberDao memberDao) {
+        this.memberDao = memberDao;
+    }
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.getParameterType().equals(Member.class);
+    }
+
+    @Override
+    public Object resolveArgument(
+            MethodParameter parameter,
+            ModelAndViewContainer mavContainer,
+            NativeWebRequest webRequest,
+            WebDataBinderFactory binderFactory
+    ) {
+        String authorization = webRequest.getHeader(HttpHeaders.AUTHORIZATION);
+        if (authorization == null) {
+            return null;
+        }
+
+        String[] authHeader = authorization.split(" ");
+        if (!authHeader[0].equalsIgnoreCase("basic")) {
+            return null;
+        }
+
+        byte[] decodedBytes = Base64.decodeBase64(authHeader[1]);
+        String decodedString = new String(decodedBytes);
+
+        String[] credentials = decodedString.split(":");
+        String email = credentials[0];
+        String password = credentials[1];
+
+        // 본인 여부 확인
+        Member member = memberDao.getMemberByEmail(email);
+        if (!member.checkPassword(password)) {
+            throw new MemberException(MISMATCH_AUTHENTICATION);
+        }
+        return member;
+    }
+}
