+++ src/main/java/cart/product/presentation/ProductApiController.java
@@ -1,0 +1,52 @@
+package cart.product.presentation;
+
+import cart.product.application.ProductService;
+import cart.product.application.dto.CreateProductCommand;
+import cart.product.presentation.dto.CreateProductRequest;
+import cart.product.presentation.dto.UpdateProductRequest;
+import cart.product.presentation.dto.ProductResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping("/products")
+public class ProductApiController {
+
+    private final ProductService productService;
+
+    public ProductApiController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ProductResponse>> getAllProducts() {
+        return ResponseEntity.ok(productService.getAllProducts());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<ProductResponse> getProductById(@PathVariable Long id) {
+        return ResponseEntity.ok(productService.getProductById(id));
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createProduct(@RequestBody CreateProductRequest request) {
+        Long id = productService.createProduct(request.toCommand());
+        return ResponseEntity.created(URI.create("/products/" + id)).build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Void> updateProduct(@PathVariable Long id, @RequestBody UpdateProductRequest request) {
+        productService.updateProduct(id, request.toCommand());
+        return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        productService.deleteProduct(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    }
+}
