+++ src/main/java/cart/product/presentation/ProductViewController.java
@@ -1,0 +1,22 @@
+package cart.product.presentation;
+
+import cart.product.application.ProductService;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ProductViewController {
+
+    private final ProductService productService;
+
+    public ProductViewController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping("/admin")
+    public String admin(Model model) {
+        model.addAttribute("products", productService.getAllProducts());
+        return "admin";
+    }
+}
